vendor-new: version actualizada del sistema sobre la base
Representa la version nueva (pro8-main-actualizado) superpuesta sobre el
commit base fa37c10 para servir de rama de merge de 3 vias.

+++ database/migrations/tenant/2025_12_08_000002_create_item_modifier_group_table.php
@@ -15,9 +15,9 @@ class CreateItemModifierGroupTable extends Migration
     {
         if (! Schema::hasTable('item_modifier_group')) {
             Schema::create('item_modifier_group', function (Blueprint $table) {
-                $table->bigIncrements('id');
+                $table->increments('id');
                 $table->unsignedInteger('item_id');
-                $table->unsignedBigInteger('modifier_group_id');
+                $table->unsignedInteger('modifier_group_id');
                 $table->boolean('default_open')->default(false);
                 $table->timestamps();
